build: Make main project recompile when library files change
When something in a library changes, the main project should be recompiled.

But there's a bug:
Because of the -j MAKEFLAG, multiple make jobs get executed at the same time. This means the library files might get checked for changes before the library actually compiles any changes.
In that case, make has to be run again to include the updated library into the main executable.