refactor: Refactor expansion in executor
- Now that the the cmd_name and cmd_args are merged, the whole of it can be expanded into a list and then converted back to an array.
This make the code much shorter, more efficient, and easier to expand for other members of the final_cmd_table that need to be expanded.

- Change the interface of the expander.
It does not modify the given string anymore.
It can now easier communicate different outcomes of the expansion by changing the return type from bool to int.

- Always reset the static variables in the expander.

- free_array() sets the array to NULL now.
This makes it easier to avoid double frees.