feat: Find places where to `handle_assignment()`
If there is a command after the assignments, the assignments should only apply for the environment of the command during its execution.
If there is no command, only assignments, they should apply to the shell's environment.

Our process flow would already handle this with the forks, however, the same rules also apply for builtins, and they don't fork.