refactor: Refactor expander
Instead of immediatly expanding into the string, create a stack of tasks that mark the index and the length of one section important for the expander.
A stack, bc the expander has to finally expand the sequences from the back of the string to the front.
This way, the indexes don't have to be updated.

The removal of quotes coming from variables issue gets fixed bc at time of creation of the tasks, the string to be expanded does not get changed yet, and so the contents of the variables cannot disrupt.

* Resolves [BUG] Quotes that came from variable expansions should not be removed #173