style: Remove meaningless return statements
The return statements are an artifact from before the error handling functions via signals were changed.

`SIGTERM` only doesn't exit in subshell level 0.
Therefore, in any functions that are accessable from the main process, a continuation after raise_error has to be considered.
`exec_simple_cmd()` is not accessable from the main process.
And in `handle_builtin()`, there is an explicit check that this is not in subshell level 0.