build: Add ffclean goal to Makefile
It forcefully cleans up the _obj and _dep directories in the build directory.
When source filenames change, the build artifacts with the old filenames do not get deleted by fclean. This is for safety so that only files known to the Makefile get deleted, and no files that might have gotten into these directories by accident.