feat: Fake-clear the terminal at startup
It just moves the existing content upwards and starts the new content from the top of the terminal.

`\e[2J` moves the viewable lines out of view.
`\e[1;1H` moves the cursor to row 1, column 1.

This will give the illusion of clearing the screen, but it's still possible to scroll up to see the previous content.