fix: Fix interleaved error messages
If multiple processes print to stderr at the same time, with `ft_dprintf()`, multiple calls to `write()` might be called.
This can cause the error messages to be interleaved.
The solution is to first use `ft_vsnprintf()` to convert the format string with all its arguments into a buffer string, and then print that string at once with one call to `write()`.

The buffer string for the error message is set to 131072 bytes.
If an error message is longer, it would get truncated.