lib: Add `ft_vsnprintf()` to libft
C-99 Standard:
A va_list may be passed as an argument to
another function; if that function invokes the va_arg macro, the
value of the va_list in the calling function is indeterminate and shall be passed to the va_end
macro prior to any further reference.

It is permitted to create a pointer to a va_list and pass that pointer to another function, in which
case the original function may make further use of the original list after the other function returns.