fix: Change `SIGABRT` to `SIGUSR1`
While `SIGABRT` defaults to termination, it is also expected to produce a core dump. If we don't do that, we would change the expected behavior of `SIGABRT`.
An example usage that produces a `SIGABRT` signal is the `abort()` function.
Note: The `abort()` function would usually still work as expected:
> If the SIGABRT signal is ignored, or caught by a handler that returns, the abort() function will still terminate the process.  It does this by restoring the default disposition for SIGABRT and then raising the signal for a second time.
However, in our case our signal handler for `SIGABRT` does not return but exit, hence the abort function never gets the chance to send the signal a second time with default behavior.