refactor: Exit for SIGPIPE signal without leaks
By setting up a signal handler that cleans and exits for SIGPIPE, we exit with the first `write()` to a broken pipe without trying over and over again.
To avoid memory leaks, allocations in builtins that have not been freed before a call to `write()` have to be safed into the `shell` struct so that they can be freed in the signal handler.