ci: Call mstest in compiled sigpipe_wrapper to reset SIGPIPE handler
This is a workaround for the default GitHub Actions runner which inherits to all jobs and steps a SIGPIPE handler that does nothing.
The behavior can be reproduced locally by running `trap '' SIGPIPE` and then commands that produce a SIGPIPE signal, like `cat | ls`.

The workaround is to compile a C program that resets the SIGPIPE handler and call the mstest command from it.
A shell script wrapper like `trap - SIGPIPE ; exec` does not work.

Reference about the behavior of GitHub Actions runner:
https://github.com/actions/runner/issues/2684