style: Rename "bracket" to "parentheses"
- Parentheses: `( )`
- Brackets: `[ ]`
- Braces: `{ }`

All of them mean something different for bash ofc.
In bash's documentation they use the same terminology.

Leaving it as it is with bracket is confusing when going through the code and you start to wonder if we handle `[ ]`, which we don't.

I guess you can name all of them brackets, but since we have support for both parentheses `( )` and braces `{ }`, using the term in a general way is just not very clear.