fix: Fix issues with emoji in prompt
Once the color escape sequences caused no issues anymore, there were sometimes still small issues.
'🌊' is a 4-byte character, but only takes up width of 2 characters, which caused those remaining issues with readline.
The solution is to use hexcode representation of '🌊' to be able to mark 2 out of the 4 bytes as invisible characters for readline.

I found out about this workaround here: https://stackoverflow.com/a/56688167/24880406