ci: Fix test case printout
The issue was caused by a change to the tester.
The tester now uses `\033[0m` instead of `\033[m` to reset the colors.
I need to get the line number to print the test case, and the script detected the line number to be purely the last number, which now was always 0.
When passing 0 to another `sed` call this error happens.