refactor: Process new env node fully before adding it to env list
Before, the new env node was added to the env list directly after it was created, and then possibly removed again by the functions checking for special env variables.
Now, those functions operate only on the new env node and don't have to modify the env list.
`ft_lstadd_back_tail()` does nothing if the node to add is `NULL`.